+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v7.1.0
+## Added
+- [ISPI-2644] New Activity with new payment flow that use payment methods as strings
+
 # v7.0.0
 ## Added
 - [ISPI-2848] Bump integration SDK to 7.0.0